% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_to_column.R
\name{apply_to_column}
\alias{apply_to_column}
\title{apply_to_column: Apply a specified function to selected columns of a CSV file.}
\usage{
apply_to_column(file_path, function_name, columns, include_columns = TRUE)
}
\arguments{
\item{file_path}{A character string containing the file path to a CSV file to be processed.}

\item{function_name}{A function to be applied to the selected columns.}

\item{columns}{A character vector of column names to include or exclude from processing.}

\item{include_columns}{A logical value indicating whether to include (TRUE) or exclude (FALSE) the specified columns.}
}
\value{
A table containing the results of applying the specified function to the selected columns.
}
\description{
apply_to_column: Apply a specified function to selected columns of a CSV file.
}
\examples{
do_application("/home/caspian-lilypad/Downloads/move/summer 2025/capm_edited - capm_S25_fixed.csv",mean,c("Date"),TRUE)
result:
          BB         BBMR         TSXC       TSXCMR         GVTB       GVTBMR
7.368983e+00 1.301547e-02 2.052131e+04 9.028865e-03 2.487576e+00 2.045685e-03
          XT           YT
6.983180e-03 1.096979e-02



}
